chore: renaming const to primitives
